+++ settings.py
@@ -13,3 +13,11 @@ CONSISTENCY_TABLE_ID_IGNORED_RANGE = []
 CONN_ERR_MAX_RETRIES = 3
 CONN_ERR_MIN_WAIT = 1  # minimum wait between iterations in seconds
 CONN_ERR_MULTIPLIER = 2  # multiplier for the accumulated wait on each iteration
+
+# Minimum consistency check verdict interval for start to consider inconsistencies.
+CONSISTENCY_MIN_VERDICT_INTERVAL = 60 * 2
+"""
+Consistency check is eventually consistent, so the minimum interval is recommended
+to be at least greater than FLOW_STATS and ideally it slightly greater than
+whichever longest network convergence FlowMods operations that your network has.
+"""