+++ tests/unit/test_flow_controller.py
@@ -41,6 +41,7 @@ class TestFlowController(TestCase):  # pylint: disable=too-many-public-methods
                     ("flow.cookie", 1),
                     ("state", 1),
                     ("inserted_at", 1),
+                    ("updated_at", 1),
                 ],
             ),
             (
@@ -67,25 +68,12 @@ class TestFlowController(TestCase):  # pylint: disable=too-many-public-methods
         args = self.flow_controller.db.flows.bulk_write.call_args[0]
         assert len(args[0]) == len(flow_dicts)
 
-    def test_update_flow_state(self) -> None:
-        """Test update_flow_state."""
-        assert self.flow_controller.update_flow_state(self.flow_id, "installed")
-        arg1, arg2 = self.flow_controller.db.flows.find_one_and_update.call_args[0]
-        assert arg1 == {"flow_id": self.flow_id}
-        assert arg2["$set"]["state"] == "installed"
-
     def test_update_flows_state(self) -> None:
         """Test update_flows_state."""
         assert self.flow_controller.update_flows_state([self.flow_id], "installed")
         arg1, arg2 = self.flow_controller.db.flows.update_many.call_args[0]
         assert arg1 == {"flow_id": {"$in": [self.flow_id]}}
         assert arg2["$set"]["state"] == "installed"
-
-    def test_delete_flows_by_ids(self) -> None:
-        """Test delete_flows_by_ids."""
-        assert self.flow_controller.delete_flows_by_ids([self.flow_id])
-        args = self.flow_controller.db.flows.delete_many.call_args[0]
-        assert args[0] == {"flow_id": {"$in": [self.flow_id]}}
 
     def test_delete_flow_by_id(self) -> None:
         """Test delete_flow_by_id."""
@@ -101,6 +89,7 @@ class TestFlowController(TestCase):  # pylint: disable=too-many-public-methods
         args = self.flow_controller.db.flows.find.call_args[0]
         assert args[0] == {
             "updated_at": {"$lte": dt},
+            "state": {"$ne": "deleted"},
             "switch": self.dpid,
         }
         args = self.flow_controller.db.flows.find(
@@ -112,20 +101,21 @@ class TestFlowController(TestCase):  # pylint: disable=too-many-public-methods
         """Test get_flows."""
         assert not list(self.flow_controller.get_flows(self.dpid))
         args = self.flow_controller.db.flows.find.call_args[0]
-        assert args[0] == {"switch": self.dpid}
+        assert args[0] == {"switch": self.dpid, "state": {"$ne": "deleted"}}
 
-    def test_get_flows_by_cookie(self) -> None:
-        """Test get_flows_by_cookie."""
-        cookie = 0
-        assert not list(self.flow_controller.get_flows_by_cookie(self.dpid, cookie))
-        args = self.flow_controller.db.flows.find.call_args[0]
-        assert args[0] == {
-            "switch": self.dpid,
-            "flow.cookie": Decimal128(Decimal(cookie)),
+    def test_get_flows_by_cookies(self) -> None:
+        """Test get_flows_by_cookies."""
+        cookies = [0]
+        assert not list(self.flow_controller.get_flows_by_cookies([self.dpid], cookies))
+        args = self.flow_controller.db.flows.aggregate.call_args[0]
+        assert args[0][0]["$match"] == {
+            "switch": {"$in": [self.dpid]},
+            "state": {"$ne": "deleted"},
+            "flow.cookie": {"$in": [Decimal128(Decimal(cookie)) for cookie in cookies]},
         }
 
     def test_get_flows_by_state(self) -> None:
-        """Test get_flows_by_cookie."""
+        """Test get_flows_by_state."""
         state = "installed"
         assert not list(self.flow_controller.get_flows_by_state(self.dpid, state))
         args = self.flow_controller.db.flows.find.call_args[0]
@@ -138,12 +128,9 @@ class TestFlowController(TestCase):  # pylint: disable=too-many-public-methods
         assert args[0] == {"_id": self.dpid}
         assert args[1]["$set"]["state"] == "active"
 
-    def test_get_flow_check_gte_updated_at(self) -> None:
-        """Test get_flow_check_gte_updated_at."""
-        dt = datetime.utcnow()
+    def test_get_flow_check(self) -> None:
+        """Test get_flow_check."""
         state = "active"
-        assert self.flow_controller.get_flow_check_gte_updated_at(
-            self.dpid, dt, state=state
-        )
+        assert self.flow_controller.get_flow_check(self.dpid, state=state)
         args = self.flow_controller.db.flow_checks.find_one.call_args[0]
-        assert args[0] == {"_id": self.dpid, "state": state, "updated_at": {"$gte": dt}}
+        assert args[0] == {"_id": self.dpid, "state": state}